refactor(client): move infra card to correct place
