--- libs/client/infra/ui-card/src/index.ts
@@ -1,1 +1,0 @@
-export * from './lib/card.component';
