+++ libs/client/infra/ui/card/eslint.config.mjs
@@ -1,6 +1,6 @@
 import nx from '@nx/eslint-plugin';
 
-import baseConfig from '../../../../eslint.config.mjs';
+import baseConfig from '../../../../../eslint.config.mjs';
 
 export default [
   ...nx.configs['flat/angular'],
