+++ libs/client/infra/ui/card/src/index.ts
@@ -1,0 +1,1 @@
+export { CardComponent } from './lib/card.component';
