+++ libs/client/infra/ui/card/src/lib/card.component.scss
