+++ tsconfig.base.json
@@ -84,14 +84,14 @@
       "@steam-idler/client/i18n/core": ["libs/client/i18n/core/src/index.ts"],
       "@steam-idler/client/i18n/types": ["libs/client/i18n/types/src/index.ts"],
       "@steam-idler/client/i18n/ui": ["libs/client/i18n/ui/src/index.ts"],
-      "@steam-idler/client/infra/ui-card": [
-        "libs/client/infra/ui-card/src/index.ts"
-      ],
       "@steam-idler/client/accounts/data-access": [
         "libs/client/accounts/data-access/src/index.ts"
       ],
       "@steam-idler/client/accounts/feature": [
         "libs/client/accounts/feature/src/index.ts"
+      ],
+      "@steam-idler/client/infra/ui/card": [
+        "libs/client/infra/ui/card/src/index.ts"
       ]
     }
   },
